Add automatic tests on push with GitHub Actions
Also fixed an issue with bin/build not making the script executable
(which I didn't notice because it was already executable on my machine).

As a result, I added a smoke test to the build - because there's no
point running all of the tests and trying to track down the source of
the errors if it doesn't even run - which I did for several hours,
because kcov doesn't give a helpful error in that case!